+++ addons/mail/models/__init__.py
@@ -27,4 +27,5 @@ from . import ir_attachment
 from . import ir_autovacuum
 from . import ir_http
 from . import ir_model
+from . import ir_model_fields
 from . import ir_ui_view
